Move asserts to one file
