+++ src/asserts.ts
@@ -39,3 +39,15 @@ export function invalidType(v: unknown, t: string): string {
 export function throwInvalidType(v: unknown, t: string): never {
   throw new Error(invalidType(v, t));
 }
+
+export function assertNotNull<T>(v: T | null): asserts v is T {
+  if (v === null) {
+    throw new Error('Expected non-null value');
+  }
+}
+
+export function assertNotUndefined<T>(v: T | undefined): asserts v is T {
+  if (v === undefined) {
+    throw new Error('Expected non undefined value');
+  }
+}
